refactor(handlers): extract read.ts handler from intent.ts (Stage B)
handleSAPRead + version/draft resolution + BTP hints. Exports
resolveVersionAndDraftInfo + SourceVersion for the write handler (acyclic).
3672 tests, A1 zero-diff.
